refactor(ui): extract wizard step controls into a shared component
